Moved stories to src, fixed missing types
