+++ package.json
@@ -54,9 +54,9 @@
     "@storybook/addon-actions": "6.3.2",
     "@storybook/addon-docs": "6.3.7",
     "@storybook/addon-links": "6.3.7",
+    "@storybook/addon-viewport": "6.3.7",
     "@storybook/addons": "6.3.7",
     "@storybook/react": "6.3.7",
-    "@storybook/addon-viewport": "6.3.7",
     "babel-loader": "^8.2.2",
     "husky": "^7.0.1",
     "react": "^17.0.2",
@@ -74,6 +74,7 @@
     "@chakra-ui/react": "^1.6.6",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
+    "@types/react-dom": "^17.0.9",
     "framer-motion": "^4"
   }
 }
