+++ src/stories/atoms/Text.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from '@storybook/react';
 import React from 'react';
-import { Text, TextProps } from '../../src';
+import { Text, TextProps } from '../..';
 
 
 export default {
