+++ test/Paragraph.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Paragraph } from '../stories/atoms/Text.stories';
+import { Paragraph } from '../src/stories/atoms/Text.stories';
 
 const content: string =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.';
@@ -8,7 +8,7 @@ const content: string =
 describe('ButtonGroup', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<Paragraph content={content} />, div);
+    ReactDOM.render(<Paragraph variant="textOne" content={content} />, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 });
