+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-  "include": ["src", "types", "./stories", "./test"],
+  "include": ["src", "types", "src/stories", "./test"],
   "compilerOptions": {
     "module": "esnext",
     "lib": ["dom", "esnext"],
